+++ Machines/Electron/Electron.hpp
@@ -58,9 +58,6 @@ class Machine:
 			is enabled — it acts as if it were sideways RAM. Otherwise the slot is modelled as containing ROM.
 		*/
 		virtual void set_rom(ROMSlot slot, const std::vector<uint8_t> &data, bool is_writeable) = 0;
-
-		/// Enables or disables turbo-speed tape loading.
-		virtual void set_use_fast_tape_hack(bool activate) = 0;
 };
 
 }
